# docs/swagger.yaml
basePath: /
definitions:
  github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: Success
        type: string
      meta:
        $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.PaginationMeta'
      status:
        example: success
        type: string
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.Assignment:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      creator:
        $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.User'
      end_date:
        type: string
      id:
        type: string
      location:
        $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.Location'
      location_id:
        type: string
      notes:
        type: string
      personnel:
        allOf:
        - $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.Personnel'
        description: Relations
      personnel_id:
        type: string
      shift:
        $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.Shift'
      shift_id:
        type: string
      start_date:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.Bujp:
    properties:
      address:
        type: string
      bjb_account_number:
        type: string
      code:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      locations:
        items:
          $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.Location'
        type: array
      name:
        type: string
      personnels:
        items:
          $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.Personnel'
        type: array
      phone:
        type: string
      pic_name:
        type: string
      pic_phone:
        type: string
      pks_document:
        type: string
      pks_number:
        type: string
      pks_validity_end:
        type: string
      pks_validity_start:
        type: string
      shifts:
        items:
          $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.Shift'
        type: array
      status:
        type: string
      updated_at:
        type: string
      users:
        description: Relations
        items:
          $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.User'
        type: array
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.CreateAssignmentRequest:
    properties:
      created_by:
        type: string
      end_date:
        description: 'format: 2006-01-02'
        type: string
      location_id:
        type: string
      notes:
        type: string
      personnel_id:
        type: string
      shift_id:
        type: string
      start_date:
        description: 'format: 2006-01-02'
        type: string
      status:
        enum:
        - active
        - completed
        - cancelled
        type: string
    required:
    - location_id
    - personnel_id
    - start_date
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.Location:
    properties:
      address:
        type: string
      attendance_radius:
        type: integer
      bujp:
        allOf:
        - $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.Bujp'
        description: Relations
      bujp_id:
        type: string
      code:
        type: string
      created_at:
        type: string
      guards_needed:
        type: integer
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      pic_name:
        type: string
      pic_phone:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.PaginationMeta:
    properties:
      has_next:
        example: true
        type: boolean
      has_previous:
        example: false
        type: boolean
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.Personnel:
    properties:
      account_number:
        type: string
      address:
        type: string
      bank_name:
        type: string
      base_salary:
        type: number
      birth_date:
        type: string
      bujp:
        allOf:
        - $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.Bujp'
        description: Relations
      bujp_id:
        type: string
      contract_end_date:
        type: string
      created_at:
        type: string
      email:
        type: string
      emergency_phone:
        type: string
      full_name:
        type: string
      gender:
        type: string
      id:
        type: string
      id_number:
        type: string
      join_date:
        type: string
      license_number:
        type: string
      mother_maiden_name:
        type: string
      phone:
        type: string
      photo:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.User'
      user_id:
        type: string
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.Shift:
    properties:
      bujp:
        allOf:
        - $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.Bujp'
        description: Relations
      bujp_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      duration_hours:
        type: number
      end_time:
        $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.TimeOnly'
      id:
        type: string
      name:
        type: string
      start_time:
        $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.TimeOnly'
      status:
        type: string
      updated_at:
        type: string
    type: object
  github_com_ganiramadhan_ganipedia_backend_internal_model.TimeOnly:
    properties:
      time.Time:
        type: string
    type: object
  github_com_ganiramadhan_ganipedia_backend_internal_model.UpdateAssignmentRequest:
    properties:
      end_date:
        description: 'format: 2006-01-02'
        type: string
      location_id:
        type: string
      notes:
        type: string
      shift_id:
        type: string
      start_date:
        description: 'format: 2006-01-02'
        type: string
      status:
        enum:
        - active
        - completed
        - cancelled
        type: string
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.User:
    properties:
      bujp:
        $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.Bujp'
      bujp_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      last_login:
        type: string
      phone:
        type: string
      photo:
        type: string
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@hrmis.com
    name: API Support
  description: HRMIS Backend API - Human Resource Management Information System
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: HRMIS API
  version: "1.0"
paths:
...
  /assignments:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by personnel ID
        in: query
        name: personnel_id
        type: string
      - description: Filter by location ID
        in: query
        name: location_id
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.Assignment'
                  type: array
                meta:
                  $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.PaginationMeta'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all assignments
      tags:
      - Assignments
    post:
      parameters:
      - description: Assignment data
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.CreateAssignmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.Assignment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Create new assignment
      tags:
      - Assignments
  /assignments/{id}:
    delete:
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete assignment
      tags:
      - Assignments
    get:
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.Assignment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Get assignment by ID
      tags:
      - Assignments
    put:
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: string
      - description: Assignment data
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.UpdateAssignmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.Assignment'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Update assignment
      tags:
      - Assignments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"